% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagouv-package.R
\docType{package}
\name{datagouv-package}
\alias{datagouv}
\alias{datagouv-package}
\title{datagouv: Tools to download and explore datasets from the French open data
platform}
\description{
This package provides a small client for the public API of
\href{https://www.data.gouv.fr}{data.gouv.fr}, the French government's open data
platform. It lets you list the published datasets, download their tabular
resources and compute summary metrics (file weight, number of columns,
missing-value rates, ...). Requests are built on top of the \code{httr2} package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stamm-a.github.io/datagouv/}
  \item Report bugs at \url{https://github.com/stamm-a/datagouv/issues}
}

}
\author{
\strong{Maintainer}: A. Stamm \email{first.last@example.com}

Authors:
\itemize{
  \item A. Stamm \email{first.last@example.com}
}

}
\keyword{internal}
